update ver 0.09

+++ SYSTEM/SAVEDATA/savedata_delete.rb
@@ -5,29 +5,40 @@
 #セーブデータ１を削除
 def savedata1_delete
 
-	#セーブデータ１の削除
-	File.delete("DATA/SAVE/savedata1_choice.txt")
-	File.delete("DATA/SAVE/savedata1_lineno.txt")
-	File.delete("DATA/SAVE/savedata1.png")
+	savedata_exist = savedata1_check()
+	
+	if savedata_exist == 1 then
+		#セーブデータ１の削除
+		File.delete("DATA/SAVE/savedata1_choice.txt")
+		File.delete("DATA/SAVE/savedata1_lineno.txt")
+		File.delete("DATA/SAVE/savedata1.png")
+	end
 end
 
 #セーブデータ２を削除
 def savedata2_delete
 
-	#セーブデータ２の削除
-	File.delete("DATA/SAVE/savedata2_choice.txt")
-	File.delete("DATA/SAVE/savedata2_lineno.txt")
-	File.delete("DATA/SAVE/savedata2.png")
+	savedata_exist = savedata2_check()
+	
+	if savedata_exist == 1 then
+		#セーブデータ２の削除
+		File.delete("DATA/SAVE/savedata2_choice.txt")
+		File.delete("DATA/SAVE/savedata2_lineno.txt")
+		File.delete("DATA/SAVE/savedata2.png")
+	end
 end
 
 #セーブデータ３を削除
 def savedata3_delete
 
-	#セーブデータ３の削除
-	File.delete("DATA/SAVE/savedata3_choice.txt")
-	File.delete("DATA/SAVE/savedata3_lineno.txt")
-	File.delete("DATA/SAVE/savedata3.png")
-
+	savedata_exist = savedata3_check()
+	
+	if savedata_exist == 1 then
+		#セーブデータ３の削除
+		File.delete("DATA/SAVE/savedata3_choice.txt")
+		File.delete("DATA/SAVE/savedata3_lineno.txt")
+		File.delete("DATA/SAVE/savedata3.png")
+	end
 end
 
 #savedata_saveメソッドの定義
@@ -56,38 +67,11 @@ def savedata_delete
 		#カーソルの描画
 		Window.draw_font(50, menu_y, "□", font)
 
-		if menu_y == 300 then
-			#セーブデータ画像の描画
-			Window.draw_scale(100, -100, savedata1_screenshot, 0.5, 0.5, 1)
-		end
+		#各種スクリーンショット描画
+		savedata_screenshot_draw(savedata1_screenshot, savedata2_screenshot, savedata3_screenshot, menu_y)
 
-		if menu_y == 350 then
-			#セーブデータ画像の描画
-			Window.draw_scale(100, -100, savedata2_screenshot, 0.5, 0.5, 1)
-		end
-
-		if menu_y == 400 then
-			#セーブデータ画像の描画
-			Window.draw_scale(100, -100, savedata3_screenshot, 0.5, 0.5, 1)
-		end
-
-		#カーソルのキー操作（↑キー）
-		if Input.key_push?(K_UP) then
-			menu_y -= 50
-
-			if menu_y <= 250 then
-				menu_y = 450
-			end
-		end
-
-		#カーソルのキー操作（↓キー）
-		if Input.key_push?(K_DOWN) then
-			menu_y += 50
-
-			if menu_y >= 500 then
-				menu_y = 300
-			end
-		end
+		#キー操作関連
+		menu_y = savedata_key(menu_y)
 
 		#「セーブデータ１にセーブ」を選択した際の処理
 		if Input.key_push?(K_RETURN) && menu_y == 300 then
