+++ SYSTEM/SAVEDATA/savedata_load.rb
@@ -5,34 +5,44 @@
 #セーブデータ１のロード
 def savedata1_load
 
-	#ルート管理変数への代入（セーブデータの読み込み)
-	$choice = File.open("DATA/SAVE/savedata1_choice.txt", "r").read.to_i
 
-	#スクリプト管理変数への代入（セーブデータの読み込み)
-	$lineno = File.open("DATA/SAVE/savedata1_lineno.txt", "r").read.to_i
+	savedata_exist = savedata1_check()
+	
+	if savedata_exist == 1 then
+		#ルート管理変数への代入（セーブデータの読み込み)
+		$choice = File.open("DATA/SAVE/savedata1_choice.txt", "r").read.to_i
 
+		#スクリプト管理変数への代入（セーブデータの読み込み)
+		$lineno = File.open("DATA/SAVE/savedata1_lineno.txt", "r").read.to_i
+	end
 end
 
 #セーブデータ２のロード
 def savedata2_load
 
-	#ルート管理変数への代入（セーブデータの読み込み)
-	$choice = File.open("DATA/SAVE/savedata2_choice.txt", "r").read.to_i
+	savedata_exist = savedata2_check()
+	
+	if savedata_exist == 1 then
+		#ルート管理変数への代入（セーブデータの読み込み)
+		$choice = File.open("DATA/SAVE/savedata2_choice.txt", "r").read.to_i
 
-	#スクリプト管理変数への代入（セーブデータの読み込み)
-	$lineno = File.open("DATA/SAVE/savedata2_lineno.txt", "r").read.to_i
-
+		#スクリプト管理変数への代入（セーブデータの読み込み)
+		$lineno = File.open("DATA/SAVE/savedata2_lineno.txt", "r").read.to_i
+	end
 end
 
 #セーブデータ３のロード
 def savedata3_load
 
-	#ルート管理変数への代入（セーブデータの読み込み)
-	$choice = File.open("DATA/SAVE/savedata3_choice.txt", "r").read.to_i
+	savedata_exist = savedata3_check()
+	
+	if savedata_exist == 1 then
+		#ルート管理変数への代入（セーブデータの読み込み)
+		$choice = File.open("DATA/SAVE/savedata3_choice.txt", "r").read.to_i
 
-	#スクリプト管理変数への代入（セーブデータの読み込み)
-	$lineno = File.open("DATA/SAVE/savedata3_lineno.txt", "r").read.to_i
-
+		#スクリプト管理変数への代入（セーブデータの読み込み)
+		$lineno = File.open("DATA/SAVE/savedata3_lineno.txt", "r").read.to_i
+	end
 end
 
 #savedata_loadメソッドの定義
@@ -61,41 +71,11 @@ def savedata_load
 		#カーソルの描画
 		Window.draw_font(50, menu_y, "□", font)
 
-		if menu_y == 300 then
+		#各種スクリーンショット描画
+		savedata_screenshot_draw(savedata1_screenshot, savedata2_screenshot, savedata3_screenshot, menu_y)
 
-			#セーブデータ画像の描画
-			Window.draw_scale(100, -100, savedata1_screenshot, 0.5, 0.5, 1)
-		end
-
-		if menu_y == 350 then
-
-			#セーブデータ画像の描画
-			Window.draw_scale(100, -100, savedata2_screenshot, 0.5, 0.5, 1)
-		end
-
-		if menu_y == 400 then
-
-			#セーブデータ画像の描画
-			Window.draw_scale(100, -100, savedata3_screenshot, 0.5, 0.5, 1)
-		end
-
-		#カーソルのキー操作（↑キー）
-		if Input.key_push?(K_UP) then
-			menu_y -= 50
-
-			if menu_y <= 250 then
-				menu_y = 450
-			end
-		end
-
-		#カーソルのキー操作（↓キー）
-		if Input.key_push?(K_DOWN) then
-			menu_y += 50
-
-			if menu_y >= 500 then
-				menu_y = 300
-			end
-		end
+		#キー操作関連
+		menu_y = savedata_key(menu_y)
 
 		#「セーブデータ１をロード」を選択した際の処理
 		if Input.key_push?(K_RETURN) && menu_y == 300 then
